requirements.txt
Django==1.8.2
django-extensions==1.5.5
djangorestframework==3.1.3
six==1.9.0
Werkzeug==0.10.4
